+++ camp44/crud/__init__.py
@@ -1,1 +1,1 @@
-from . import app, user, entity
+from . import app, user, entity, refresh_token